+++ ops/ansible/ansible_install.sh
@@ -1,2 +1,3 @@
 #!/bin/bash
+pip install --upgrade ansible setuptools
 ansible-playbook -i "localhost," -c local /opt/postmaster/git/ops/ansible/site.yml