+++ ops/ansible/ansible_upgrade.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+pip install ansible
+ansible-playbook -i "localhost," -c local /opt/postmaster/git/ops/ansible/upgrade.yml