+++ docs/ChangeLog.md
@@ -10,11 +10,13 @@ Features:
 
 Improvements:
 
+* **New Command**: `python manage.py existingdb` for database migrations for existing mailserver database configuration before PostMaster [GH-138]
 * Added additional documentation [GH-115]
 * Improved Active Directory authentication performance [GH-118]
 * Cleaned up JavaScript event listeners [GH-120]
 * Vagrant now uses Ansible for configuration instead of a bash script [GH-111]
 * Updated the Python packages to the latest versions [GH-127]
+* Ansible can now run proper database migrations on upgrade [GH-138]
 
 Bug Fixes: